Move functions from g_sample to sample
