+++ tests/gbs/test_g_sample.py
@@ -18,6 +18,7 @@ Unit tests for strawberryfields.gbs.graph.sample
 import networkx as nx
 import pytest
 
+import strawberryfields.gbs.sample
 from strawberryfields.gbs import g_sample
 
 pytestmark = pytest.mark.gbs
@@ -56,7 +57,9 @@ def test_sample_subgraphs_invalid_distribution(graph):
     """Tests if function ``sample.sample_subgraphs`` raises a ``ValueError`` for an
     invalid sampling distribution"""
     with pytest.raises(ValueError, match="Invalid distribution selected"):
-        g_sample.sample_subgraphs(graph, nodes=2, samples=10, sample_options={"distribution": ""})
+        strawberryfields.gbs.sample.sample_subgraphs(
+            graph, nodes=2, samples=10, sample_options={"distribution": ""}
+        )
 
 
 @pytest.mark.parametrize(
@@ -68,7 +71,7 @@ def test_sample_subgraphs_integration(graph, nodes, samples, distribution):
 
     graph = nx.relabel_nodes(graph, lambda x: x ** 2)
     graph_nodes = set(graph.nodes)
-    output_samples = g_sample.sample_subgraphs(
+    output_samples = strawberryfields.gbs.sample.sample_subgraphs(
         graph=graph, nodes=nodes, samples=samples, sample_options={"distribution": distribution}
     )
 
@@ -83,7 +86,7 @@ class TestToSubgraphs:
     def test_graph(self, graph):
         """Test if function returns correctly processed subgraphs given input samples of the list
         ``quantum_samples``."""
-        assert g_sample.to_subgraphs(graph, samples=quantum_samples) == subgraphs
+        assert strawberryfields.gbs.sample.to_subgraphs(graph, samples=quantum_samples) == subgraphs
 
     def test_graph_mapped(self, graph):
         """Test if function returns correctly processed subgraphs given input samples of the list
@@ -94,4 +97,7 @@ class TestToSubgraphs:
         graph_nodes = list(graph.nodes)
         subgraphs_mapped = [sorted([graph_nodes[i] for i in subgraph]) for subgraph in subgraphs]
 
-        assert g_sample.to_subgraphs(graph, samples=quantum_samples) == subgraphs_mapped
+        assert (
+            strawberryfields.gbs.sample.to_subgraphs(graph, samples=quantum_samples)
+            == subgraphs_mapped
+        )
